Continued Development of AI System and Action Managers
Added Priorities so certain managers start with a higher priority and
refresh timers in miliseconds so the managers don't get OCD checking the
command queue.  Added calls to the refreshStatus() methods from
Post-Int.  This will eventually be moved to the game loop when
developed, and they can be set to respond to triggers.